+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/AuthDomainGroupApiSpec.scala
@@ -3,7 +3,8 @@ package org.broadinstitute.dsde.test.api
 import cats.implicits.catsSyntaxOptionId
 import org.broadinstitute.dsde.rawls.model.WorkspaceJsonSupport._
 import org.broadinstitute.dsde.rawls.model.WorkspaceResponse
-import org.broadinstitute.dsde.test.api.tagannotation.AuthDomainsTest
+import org.broadinstitute.dsde.test.api.tagannotation.AuthDomainsTest1
+import org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestGroupAnnotations.AuthDomainsTest
 import org.broadinstitute.dsde.test.util.AuthDomainMatcher
 import org.broadinstitute.dsde.workbench.auth.AuthToken
 import org.broadinstitute.dsde.workbench.auth.AuthTokenScopes.{billingScopes, serviceAccountScopes}
