--- automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/AuthDomainsTest.java
@@ -1,4 +1,0 @@
-package org.broadinstitute.dsde.test.api.tagannotation;
-
-@RawlsTestGroupAnnotation
-public @interface AuthDomainsTest {}
