--- automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/DataRepoSnapshotsTest.java
@@ -1,5 +1,0 @@
-package org.broadinstitute.dsde.test.api.tagannotation;
-
-@RawlsTestGroupAnnotation
-public @interface DataRepoSnapshotsTest {
-}
