+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/DataRepoSnapshotsTest1.java
@@ -1,0 +1,7 @@
+package org.broadinstitute.dsde.test.api.tagannotation;
+
+import org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestAnnotationBase;
+
+@RawlsTestAnnotationBase
+public @interface DataRepoSnapshotsTest1 {
+}
