+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/MethodsTest1.java
@@ -1,0 +1,7 @@
+package org.broadinstitute.dsde.test.api.tagannotation;
+
+import org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestAnnotationBase;
+
+@RawlsTestAnnotationBase
+public @interface MethodsTest1 {
+}
