--- automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/RawlsTest.java
@@ -1,4 +1,0 @@
-package org.broadinstitute.dsde.test.api.tagannotation;
-
-@RawlsTestGroupAnnotation
-public @interface RawlsTest {}
