+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/RawlsTest1.java
@@ -1,0 +1,6 @@
+package org.broadinstitute.dsde.test.api.tagannotation;
+
+import org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestAnnotationBase;
+
+@RawlsTestAnnotationBase
+public @interface RawlsTest1 {}
