--- automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/RawlsTestObj.scala
@@ -1,5 +1,0 @@
-package org.broadinstitute.dsde.test.api.tagannotation
-
-import org.scalatest.Tag
-
-object RawlsTestObj extends Tag("org.broadinstitute.dsde.test.api.tagannotation.RawlsTest")
