--- automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/WorkspacesTest.java
@@ -1,4 +1,0 @@
-package org.broadinstitute.dsde.test.api.tagannotation;
-
-@RawlsTestGroupAnnotation
-public @interface WorkspacesTest {}
