+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/WorkspacesTestObj1.scala
@@ -1,0 +1,5 @@
+package org.broadinstitute.dsde.test.api.tagannotation
+
+import org.scalatest.Tag
+
+object WorkspacesTestObj1 extends Tag("org.broadinstitute.dsde.test.api.tagannotation.WorkspacesTestObj1")
