+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/rawls/RawlsTestGroupAnnotations.java
@@ -1,0 +1,21 @@
+package org.broadinstitute.dsde.test.api.tagannotation.rawls;
+
+public interface RawlsTestGroupAnnotations {
+    @RawlsTestAnnotationBase
+    public @interface AuthDomainsTest {}
+
+    @RawlsTestAnnotationBase
+    public @interface BillingsTest {}
+
+    @RawlsTestAnnotationBase
+    public @interface DataRepoSnapshotsTest {}
+
+    @RawlsTestAnnotationBase
+    public @interface MethodsTest {}
+
+    @RawlsTestAnnotationBase
+    public @interface RawlsTest {}
+
+    @RawlsTestAnnotationBase
+    public @interface WorkspacesTest {}
+}
