+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/rawls/RawlsTestTaggedAs.scala
@@ -1,0 +1,7 @@
+package org.broadinstitute.dsde.test.api.tagannotation.rawls
+
+import org.scalatest.Tag
+
+object MethodsTestObj extends Tag("org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestGroupAnnotations.MethodsTest")
+object RawlsTestObj extends Tag("org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestGroupAnnotations.RawlsTest")
+object WorkspacesTestObj extends Tag("org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestGroupAnnotations.WorkspacesTest")
